Small fixes

+++ src/EntityFx.BenchmarkDb.Contracts/Cpu/CrystalParameters.cs
@@ -4,7 +4,7 @@
     {
         public decimal? ProcessInNm { get; set; }
 
-        public uint? TransistorsCount { get; set; }
+        public ulong? TransistorsCount { get; set; }
 
         public uint? TDP { get; set; }
